backend/requirements.txt
python-dotenv
fastapi
uvicorn[standard]
python-multipart

langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-google-genai
google-genai
filetype
langchain-groq
chromadb
pypdf
rank-bm25
sqlalchemy
psycopg[binary]
alembic
requests
